Calculate Kendall's Tau matrix then estimate latent correlation matrix

Estimation of latent correlation matrix from observed data of (possibly) mixed types (continuous/biary/truncated continuous) based on the latent Gaussian copula model.

estR(
  X,
  types = c("ter", "con"),
  method = "approx",
  nu = 0.01,
  tol = 1e-08,
  ratio = 0.9,
  showplot = FALSE
)

Arguments

X
A numeric data matrix (n by p).

types
A vector with length of p which specifies types of variables in X correspondingly. Must be one of "con" (continuous), "bin" (binary), "tru" (truncated) or "ter" (ternary). The default value is c("ter", "con") which means the first variable is ternary, second variable is continuous.

method
The calculation method of latent correlation. Either "original" method or "approx". If method = "approx", multilinear approximation method is used, which is much faster than the original method. If method = "original", optimization of the bridge inverse function is used. The default is "approx".

nu
Shrinkage parameter for correlation matrix, must be between 0 and 1, the default value is 0.01.

tol
Desired accuracy when calculating the solution of bridge function. The default value is 1e-8.

ratio
The maximum ratio of Kendall's tau and boundary to implement multilinear interpolation. The default value is 0.9.

showplot
Plot latent correlation matrix R as a heatmap.

Value

estR returns

K: Kendall Tau (Tau-a) Matrix of X (p x p)

R: Estimated latent correlation matrix of whole X (p x p)

plotR: Heatmap plot for latent correlation matrix R

References

Fan J., Liu H., Ning Y. and Zou H. (2017) "High dimensional semiparametric latent graphicalmodel for mixed data" <doi:10.1111/rssb.12168>. Yoon G., Carroll R.J. and Gaynanova I. (2020) "Sparse semiparametric canonical correlation analysis for data of mixed types" <doi:10.1093/biomet/asaa007>. Yoon G., Müller C.L., Gaynanova I. (2020) "Fast computation of latent correlations" <arXiv:2006.13875>.

Examples

# Data generation
X = GenData()$X
# Estimate latent correlation matrix with original method
R_nc_org = estR(X = X, types = c("ter", "con"), method = "original")$R
# Estimate latent correlation matrix with aprroximation method
R_nc_approx = estR(X = X, types = c("ter", "con"), method = "approx")$R
# Heatmap for latent correlation matrix.
Heatmap_R_nc_approx = estR(X = X, types = c("ter", "con"), method = "approx", showplot = TRUE)$plotR