Numerical evaluation for different estimation methods.

Speed and accuracy comparison of two different estimation methods.

evaluation(
  genfun,
  estfun_1,
  estfun_2,
  grid_list,
  nrep = 100,
  showplot = FALSE,
  cores = detectCores(),
  ...
)

Arguments

genfun

A data generation function.

estfun_1

A function for first estimation method.

estfun_2

A function for second estimation method.

grid_list

A list for grid points to be evaluated (each element of list is a vector represents ticklabels on a dimension). The number of list elements are the dimension of function inputs.

nrep

Number of replications in simulation.

showplot

Logical indicator. showplot = TRUE generates the heatmaps of output arrays. NULL if showplot = FALSE.

cores

The numbers of cores (threads) of your machine to conduct parallel computing.

...

Other inputs for data generation or estimation functions to be passed through.

Value

evaluation returns

meanAE_1: An array for mean absolute error of first estimation method.

meanAE_2: An array for mean absolute error of second estimation method.

medianAE_1: An array for median absolute error of first estimation method.

medianAE_2: An array for median absolute error of second estimation method.

maxAE_1: An array for maximum absolute error of first estimation method.

maxAE_2: An array for maximum absolute error of second estimation method.

meanAE_diff: An array for mean absolute error of difference between two estimations.

medianAE_diff: An array for median absolute error of difference between two estimations.

maxAE_diff: An array for maximum absolute error of difference between two estimations.

mediantime_1: An array for median time of first estimation method.

mediantime_2: An array for median time of second estimation method.

plot_meanAE_1: A plot for mean absolute error of first estimation method.

plot_meanAE_2: A plot for mean absolute error of second estimation method.

plot_medianAE_1: A plot for median absolute error of first estimation method.

plot_medianAE_2: A plot for median absolute error of second estimation method.

plot_maxAE_1: A plot for maximum absolute error of first estimation method.

plot_maxAE_2: A plot for maximum absolute error of second estimation method.

plot_meanAE_diff: A plot for mean absolute error of difference between two estimations.

plot_medianAE_diff: A plot for median absolute error of difference between two estimations.

plot_maxAE_diff: A plot for maximum absolute error of difference between two estimations.

plot_mediantime_1: A plot for median time of first estimation method.

plot_mediantime_2: A plot for median time of second estimation method.

Examples

library(latentcor)
grid_list = list(LatentR = seq(-0.9, 0.9, by = 0.3), TruncRate = seq(0.1, 0.9, by = 0.3))
genfun = function (grid_input, ...) {
  out = gen_data(rhos = grid_input[1], XP = list(grid_input[2], NA), ...)$X
  return(out)
}
estfun_1 = function(X, ...) {
  out = latentcor(X = X, method = "original", ...)$R[1, 2]
  return(out)
}
estfun_2 = function(X, ...) {
  out = latentcor(X = X, method = "approx", ...)$R[1, 2]
  return(out)
}
evaluation_BC = evaluation(genfun = genfun, estfun_1 = estfun_1, estfun_2 = estfun_2,
grid_list = grid_list, nrep = 3, cores = 80, showplot = TRUE, types = c("bin", "con"))
#> Loading required package: foreach
#> Loading required package: future
#> Warning: package 'future' was built under R version 4.2.1
#> Loading required package: rngtools
#> Warning: `gather_()` was deprecated in tidyr 1.2.0.
#> Please use `gather()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
evaluation_BC$meanAE_diff
#>              [,1]         [,2]         [,3]
#> [1,] 0.000000e+00 6.400770e-05 1.174966e-04
#> [2,] 1.953253e-04 2.215685e-04 2.985310e-04
#> [3,] 6.020488e-05 2.318909e-04 2.282316e-04
#> [4,] 7.560983e-05 6.627385e-05 9.298600e-05
#> [5,] 1.218419e-04 1.655010e-04 2.043086e-04
#> [6,] 2.006572e-04 2.133672e-04 2.775776e-04
#> [7,] 0.000000e+00 7.034946e-05 8.124977e-05
evaluation_BC$plot_meanAE_diff