Estimate latent correlation for mixed types.

Estimation of latent correlation matrix from observed data of (possibly) mixed types (continuous/binary/truncated/ternary) based on the latent Gaussian copula model. Missing values (NA) are allowed. The estimation is based on pairwise complete observations.

estR(
  X,
  types = "con",
  method = c("approx", "original"),
  nu = 0.01,
  tol = 1e-08,
  ratio = 0.9,
  showplot = FALSE
)

Arguments

X
A numeric data matrix (n by p), where n is number of samples, and p is number of variables. Missing values (NA) are allowed, in which case the estimation is based on pairwise complete observations.

types
A vector of length p indicating the type of each of the p variables in X. Each element must be one of "con" (continuous), "bin" (binary), "tru" (truncated) or "ter" (ternary). If the vector has length 1, then all p variables are assumed to be of the same type that is supplied. The default value is "con" which means all variables are continuous.

method
The calculation method for latent correlations. Either "original" or "approx". If method = "approx", multilinear approximation method is used, which is much faster than the original method. If method = "original", optimization of the bridge inverse function is used. The default is "approx".

nu
Shrinkage parameter for the correlation matrix, must be between 0 and 1, the default value is 0.01.

tol
Desired accuracy when calculating the solution of bridge function. The default value is 1e-8.

ratio
The boundary value for multilinear interpolation, must be between 0 and 1, the default value is 0.9. If method = "original", this parameter is ignored.

showplot
Logical indicator. showplot = TRUE generates a ggplot object plotR with the heatmap of latent correlation matrix R. plotR = NULL if showplot = FALSE.

Value

estR returns

zratios: A list of of length p corresponding to each variable. Returns NA for continuous variable; proportion of zeros for binary/truncated variables; the cumulative proportions of zeros and ones (e.g. first value is proportion of zeros, second value is proportion of zeros and ones) for ternary variable.

K: (p x p) Kendall Tau (Tau-a) Matrix for X

R: (p x p) Estimated latent correlation matrix for X

plotR: Heatmap plot of latent correlation matrix R, NULL if showplot = FALSE

References

Fan J., Liu H., Ning Y. and Zou H. (2017) "High dimensional semiparametric latent graphicalmodel for mixed data" doi: 10.1111/rssb.12168 .

Yoon G., Carroll R.J. and Gaynanova I. (2020) "Sparse semiparametric canonical correlation analysis for data of mixed types" doi: 10.1093/biomet/asaa007 .

Yoon G., Müller C.L., Gaynanova I. (2020) "Fast computation of latent correlations" doi: 10.1080/10618600.2021.1882468 .

Examples

# Data generation
X = GenData()$X
# Estimate latent correlation matrix with original method
R_nc_org = estR(X = X, types = c("ter", "con"), method = "original")$R
# Estimate latent correlation matrix with aprroximation method
R_nc_approx = estR(X = X, types = c("ter", "con"), method = "approx")$R
# Heatmap for latent correlation matrix.
Heatmap_R_nc_approx = estR(X = X, types = c("ter", "con"), method = "approx", showplot = TRUE)$plotR