Numerical evaluation for different estimation methods.

Speed and accuracy comparison of two different estimation methods.

evaluation(
  genfun,
  estfun_1,
  estfun_2,
  grid_list,
  nrep = 100,
  showplot = FALSE,
  cores = detectCores(),
  ...
)

Arguments

genfun

A data generation function.

estfun_1

A function for first estimation method.

estfun_2

A function for second estimation method.

grid_list

A list for grid points to be evaluated (each element of list is a vector represents ticklabels on a dimension). The number of list elements are the dimension of function inputs.

nrep

Number of replications in simulation.

showplot

Logical indicator. showplot = TRUE generates the heatmaps of output arrays. NULL if showplot = FALSE.

cores

The numbers of cores (threads) of your machine to conduct parallel computing.

...

Other inputs for data generation or estimation functions to be passed through.

Value

evaluation returns

meanAE_1: An array for mean absolute error of first estimation method.

meanAE_2: An array for mean absolute error of second estimation method.

medianAE_1: An array for median absolute error of first estimation method.

medianAE_2: An array for median absolute error of second estimation method.

maxAE_1: An array for maximum absolute error of first estimation method.

maxAE_2: An array for maximum absolute error of second estimation method.

meanAE_diff: An array for mean absolute error of difference between two estimations.

medianAE_diff: An array for median absolute error of difference between two estimations.

maxAE_diff: An array for maximum absolute error of difference between two estimations.

mediantime_1: An array for median time of first estimation method.

mediantime_2: An array for median time of second estimation method.

plot_meanAE_1: A plot for mean absolute error of first estimation method.

plot_meanAE_2: A plot for mean absolute error of second estimation method.

plot_medianAE_1: A plot for median absolute error of first estimation method.

plot_medianAE_2: A plot for median absolute error of second estimation method.

plot_maxAE_1: A plot for maximum absolute error of first estimation method.

plot_maxAE_2: A plot for maximum absolute error of second estimation method.

plot_meanAE_diff: A plot for mean absolute error of difference between two estimations.

plot_medianAE_diff: A plot for median absolute error of difference between two estimations.

plot_maxAE_diff: A plot for maximum absolute error of difference between two estimations.

plot_mediantime_1: A plot for median time of first estimation method.

plot_mediantime_2: A plot for median time of second estimation method.